+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 - Delete indices when buckets and collections are deleted (fixes #21)
 - Support quick search from querystring (fixes #34)
 - Return details about invalid queries in request body (fixes #23)
+- Support defining mapping from the ``index:schema`` property in the collection metadata (ref #8)
 
 **Bug fixes**